requirements.txt
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
chromadb
python-dotenv
requests
beautifulsoup4